+++ NEWS.md
@@ -1,7 +1,9 @@
-# packcircles 0.3.6.9000 (2024-11-20)
+# packcircles 0.3.7 (2024-11-21)
 
-* Update progressive packing vignette example to use ggiraph::girafe() function
-  (thanks to David Gohel).
+* Fix: Updated progressive packing vignette example to use ggiraph::girafe() 
+  function and fix CRAN warning (thanks to David Gohel).
+  
+* Use the checkmate package to check arguments to functions.
 
 # packcircles 0.3.6 (2023-09-05)